+++ src/main/java/pe/com/ciberelectrik/restcontroller/MarcaRestController.java
@@ -14,23 +14,35 @@ public class MarcaRestController {
     private MarcaService servicio;
 
     @GetMapping
-    public List<MarcaEntity> findAll() { return servicio.findAll(); }
+    public List<MarcaEntity> findAll() {
+        return servicio.findAll();
+    }
 
     @GetMapping("/custom")
-    public List<MarcaEntity> findAllCustom() { return servicio.findAllCustom(); }
+    public List<MarcaEntity> findAllCustom() { return
+            servicio.findAllCustom();
+    }
 
     @GetMapping("/{id}")
-    public MarcaEntity findById(@PathVariable Integer id) { return servicio.findById(id); }
+    public MarcaEntity findById(@PathVariable Long id) {
+        return servicio.findById(id);
+    }
 
     @PostMapping
-    public MarcaEntity add(@RequestBody MarcaEntity obj) { return servicio.add(obj); }
+    public MarcaEntity add(@RequestBody MarcaEntity obj) {
+        return servicio.add(obj);
+    }
 
     @PutMapping("/{id}")
-    public MarcaEntity update(@RequestBody MarcaEntity obj, @PathVariable Integer id) { return servicio.update(obj, id); }
+    public MarcaEntity update(@RequestBody MarcaEntity obj, @PathVariable Long id) { return servicio.update(obj, id); }
 
     @DeleteMapping("/{id}")
-    public MarcaEntity delete(@PathVariable Integer id) { return servicio.delete(id); }
+    public MarcaEntity delete(@PathVariable Long id) {
+        return servicio.delete(id);
+    }
 
     @PutMapping("/enable/{id}")
-    public MarcaEntity enable(@PathVariable Integer id) { return servicio.enable(id); }
+    public MarcaEntity enable(@PathVariable Long id) {
+        return servicio.enable(id);
+    }
 }
